requirements.txt
asgiref==3.2.7
bson==0.5.8
certifi==2020.4.5.1
chardet==3.0.4
dataclasses==0.6
Django==2.2.12
djongo==1.3.2
idna==2.9
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.9
